Add a timeout and run tests in parallel
This adds a global timeout for all test to complete with a new "timeout"
option in the config file.
The tests are run in parallel so that the time taken by the checks does
not depend on the number of checks. The execution time will stay the
same even if new checks are added.